[Refactor] 문자열을 통한 로또 티켓 생성 책임 로또 클래스로 위임
