+++ src/main/java/lotto/LottoApplication.java
@@ -3,13 +3,12 @@ package lotto;
 import lotto.domain.Ball;
 import lotto.domain.LottoCount;
 import lotto.domain.LottoTicket;
-import lotto.domain.LottoTicketFactory;
 import lotto.domain.LottoTickets;
+import lotto.domain.LottoTicketsGenerator;
 import lotto.domain.Money;
+import lotto.domain.RandomLottoTicketGenerator;
 import lotto.domain.WinningLotto;
 import lotto.domain.WinningResult;
-import lotto.domain.LottoTicketsGenerator;
-import lotto.domain.RandomLottoTicketGenerator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -25,7 +24,7 @@ public class LottoApplication {
 			LottoTickets lottos = lottoTicketsGenerator.createLottosByCount(count);
 			OutputView.printLottos(lottos);
 
-			LottoTicket winningLottoTicket = LottoTicketFactory.of(InputView.inputWinningLotto());
+			LottoTicket winningLottoTicket = LottoTicket.of(InputView.inputWinningLotto());
 			Ball bonusBall = Ball.valueOf(InputView.inputWinningBonusBall());
 			WinningLotto winningLotto = new WinningLotto(winningLottoTicket, bonusBall);
 			WinningResult result = winningLotto.getResult(lottos);
