--- src/main/java/lotto/domain/LottoTicketFactory.java
@@ -1,17 +1,0 @@
-package lotto.domain;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import lotto.util.StringUtil;
-
-public class LottoTicketFactory {
-	public static LottoTicket of(String rawWinningLotto) {
-		List<Ball> collect = Arrays.stream(StringUtil.splitRawLottoNumbers(rawWinningLotto))
-			.mapToInt(Integer::parseInt)
-			.mapToObj(Ball::valueOf)
-			.collect(Collectors.toList());
-		return new LottoTicket(collect);
-	}
-}
